EDGOAIPMH-62 - edge common v.4.2.0

+++ src/test/java/org/folio/edge/oaipmh/utils/OaiPmhOkapiClientFactoryTest.java
@@ -12,7 +12,7 @@ import org.junit.Test;
 
 public class OaiPmhOkapiClientFactoryTest {
 
-  private static final long reqTimeout = 5000L;
+  private static final int reqTimeout = 5000;
 
   private OaiPmhOkapiClientFactory ocf;
